Move some local fixtures from CLI blockchain ursula to global fixtures
